Keep elements related to writing.
